+++ src/components/library/ImageCard.tsx
@@ -10,7 +10,7 @@ interface ImageCardProps {
 export const ImageCard = ({ image, selected, onSelect }: ImageCardProps) => {
   return (
     <div className={`image-card${selected ? ' selected' : ''}`} onClick={() => onSelect(image.id)}>
-      <img src={image.url} alt={image.filename} />
+      <img src={image.urls.thumbnail} alt={image.filename} />
       <div className="image-info">
         <span>{image.filename}</span>
         <span>{image.state?.rating ?? '\u2014'}</span>